+++ lib/engineyard/model/user.rb
@@ -1,0 +1,6 @@
+module EY
+  module Model
+    class User < ApiStruct.new(:id, :name, :email)
+    end
+  end
+end