+++ spec/spec_helper.rb
@@ -6,6 +6,7 @@ EY_ROOT = File.expand_path("../..", __FILE__)
 require 'rubygems'
 require 'bundler/setup'
 require 'escape'
+require 'net/ssh'
 
 # Bundled gems
 require 'fakeweb'